refactoring and moving files
