+++ src/components/nav/pure/link.vue
