+++ src/components/tabs/pure/item.vue
